Switch to depending on the new conda-forge libudev package for cos7.
We modify the build matrix so that two versions are build for linux64:
one using cos6 (and the libudev CDT) and one using cos7 (and the new
libudev conda package). The cos7 build will be preferred when possible
by using a higher build number. This allows us to remove not only the
dependency on the systemd CDT package but also do away with the pre- and
post- scripts (because the cos6 version will only be used when the
system is old enough to supply libudev.so.0 instead of libudev.so.1
anyway).