Add libusb-cos6 output for linux built with cos6 sysroot.
This makes it easier to select the package with the compatible
sysroot dependency when building against libusb, while users will still
get the libusb package that is compatible with their system. Now that
the conda-forge build images are cos7 but the sysroot is still cos6,
this is the easiest way to use the cos6 version compatible with the
sysroot in downstream packages which otherwise now install the cos7
version. Downstream linux/cos6 packages that have a dependency on libusb
should **add** a dependency on libusb-cos6:

requirements:
    host:
    - libusb
    - libusb-cos6  # [linux and cdt_name=='cos6']